Request a reboot instead of a poweroff
Request a reboot instead of a poweroff when exiting the init. The main
reason for the change is Firecracker, since it does not implement ACPI
and hence a shitdown will just halt the VM but the Firecracker process
will keep running. On the contrary, a reboot request will result to an
exit of the Firecracker process but also in QEMU with the respective
flags.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>