Set default route to eth0
When we boot a Linux VM and set the network configuration through
kernel boot params, we might have issues if the gateway is in a
different subnet than the IP. In particular, Linux will complain,
since it hits this point
https://elixir.bootlin.com/linux/v6.15.4/source/net/ipv4/ipconfig.c#L417

To overcome this issue, urunit will always set the default route
to be the eth0 interface.

In the future, we might need to use a different interface than eth0
or discover it.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>