Add support for reading environment variables
Add support for reading the environment variables from a config file
which is defined by the environment variable "URUNIT_CONFIG". If this
environment variable is not set, then urunit follows the previous
execution path and execs the app using the existing environment
variables.

If the environment variable sis set then it should point to the
configuration file with the environment variables. The environment
variables are in a list which starts with the special string "UES" and
ends with the special string "UEE". For instance a valid list is the
following:

UES
PATH=/bin
MYENV=myvalue
UEE

urunit will set these environment variables for the execution of the
application and in case PATH was set, it will use this environment
variable to search for the full path othe application's binary.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>