Update README with urunit config and reduce search area for app config
Add information to README regarding the new changes in urunit and
specifically for the support of a configuration file for the environment
variables and the proccess execution environment.

Furthermore, rephrase some comments in the code and reduce the search
area in the config buffer for searching the configuration for the
application execution environment.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>