Add support for reading the block mount configuration
Read from the configuration the respective info for mounting attached
block devices to the respective mountpoint. In particular the
configuration for the block mounts has the following format:

UBS
ID: <serial_id>
MP: <mountpoint>
...
UBE

Therefore, for every block device there is a pair of ID and MP where ID
refers to the serial ID of the device and MP refers to the mountpoint of
that block device.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>