Unmount all external mounts before shutdown
Make sure that all external filesystem mounts are unmounted before
shuttidng down to avoid any corruptions. External means all mounts that
use an external to the current VM source (e.g. block devices network
filesystems).

For the time being urunit checks the filesystem type in the mountinfo
and compares it with some well known block, network and cloud-based
filesystems. We might want to update this list in the future.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>